+++ packages/swapper/src/swappers/thorchain/evm/utils/getThorTxData.ts
@@ -1,10 +1,12 @@
 import { Asset } from '@shapeshiftoss/asset-service'
+import { fromAssetId } from '@shapeshiftoss/caip'
 
-import { SwapError, SwapErrorType } from '../../../../../api'
-import type { ThorchainSwapperDeps } from '../../../types'
-import { getInboundAddressDataForChain } from '../../getInboundAddressDataForChain'
-import { getLimit } from '../../getLimit/getLimit'
-import { makeSwapMemo } from '../../makeSwapMemo/makeSwapMemo'
+import { SwapError, SwapErrorType } from '../../../../api'
+import type { ThorchainSwapperDeps } from '../../types'
+import { getInboundAddressDataForChain } from '../../utils/getInboundAddressDataForChain'
+import { getLimit } from '../../utils/getLimit/getLimit'
+import { makeSwapMemo } from '../../utils/makeSwapMemo/makeSwapMemo'
+import { deposit } from '../routerCalldata'
 
 type GetBtcThorTxInfoArgs = {
   deps: ThorchainSwapperDeps
@@ -13,14 +15,14 @@ type GetBtcThorTxInfoArgs = {
   sellAmountCryptoBaseUnit: string
   slippageTolerance: string
   destinationAddress: string
-  xpub: string
   buyAssetTradeFeeUsd: string
 }
+
 type GetBtcThorTxInfoReturn = Promise<{
-  opReturnData: string
-  vault: string
-  pubkey: string
+  data: string
+  router: string
 }>
+
 type GetBtcThorTxInfo = (args: GetBtcThorTxInfoArgs) => GetBtcThorTxInfoReturn
 
 export const getThorTxInfo: GetBtcThorTxInfo = async ({
@@ -30,21 +32,18 @@ export const getThorTxInfo: GetBtcThorTxInfo = async ({
   sellAmountCryptoBaseUnit,
   slippageTolerance,
   destinationAddress,
-  xpub,
   buyAssetTradeFeeUsd,
 }) => {
   try {
-    const inboundAddress = await getInboundAddressDataForChain(
-      deps.daemonUrl,
-      sellAsset.assetId,
-      false,
-    )
+    const { assetReference, assetNamespace } = fromAssetId(sellAsset.assetId)
+    const isErc20Trade = assetNamespace === 'erc20'
+    const inboundAddress = await getInboundAddressDataForChain(deps.daemonUrl, sellAsset.assetId)
+    const router = inboundAddress?.router
     const vault = inboundAddress?.address
-
-    if (!vault)
-      throw new SwapError(`[getThorTxInfo]: vault not found for asset`, {
+    if (!inboundAddress || !router || !vault)
+      throw new SwapError(`[getPriceRatio]: inboundAddress not found for ETH`, {
         code: SwapErrorType.RESPONSE_ERROR,
-        details: { inboundAddress, sellAsset },
+        details: { inboundAddress },
       })
 
     const limit = await getLimit({
@@ -61,12 +60,15 @@ export const getThorTxInfo: GetBtcThorTxInfo = async ({
       destinationAddress,
       limit,
     })
-
-    return {
-      opReturnData: memo,
+    const data = await deposit(
+      router,
       vault,
-      pubkey: xpub,
-    }
+      isErc20Trade ? assetReference : '0x0000000000000000000000000000000000000000',
+      sellAmountCryptoBaseUnit,
+      memo,
+    )
+
+    return { data, router }
   } catch (e) {
     if (e instanceof SwapError) throw e
     throw new SwapError('[getThorTxInfo]', { cause: e, code: SwapErrorType.TRADE_QUOTE_FAILED })
