--- packages/swapper/src/swappers/thorchain/utils/evm/utils/getThorTxData.ts
@@ -1,76 +1,0 @@
-import { Asset } from '@shapeshiftoss/asset-service'
-import { fromAssetId } from '@shapeshiftoss/caip'
-
-import { SwapError, SwapErrorType } from '../../../../../api'
-import type { ThorchainSwapperDeps } from '../../../types'
-import { getInboundAddressDataForChain } from '../../getInboundAddressDataForChain'
-import { getLimit } from '../../getLimit/getLimit'
-import { makeSwapMemo } from '../../makeSwapMemo/makeSwapMemo'
-import { deposit } from '../routerCalldata'
-
-type GetBtcThorTxInfoArgs = {
-  deps: ThorchainSwapperDeps
-  sellAsset: Asset
-  buyAsset: Asset
-  sellAmountCryptoBaseUnit: string
-  slippageTolerance: string
-  destinationAddress: string
-  buyAssetTradeFeeUsd: string
-}
-
-type GetBtcThorTxInfoReturn = Promise<{
-  data: string
-  router: string
-}>
-
-type GetBtcThorTxInfo = (args: GetBtcThorTxInfoArgs) => GetBtcThorTxInfoReturn
-
-export const getThorTxInfo: GetBtcThorTxInfo = async ({
-  deps,
-  sellAsset,
-  buyAsset,
-  sellAmountCryptoBaseUnit,
-  slippageTolerance,
-  destinationAddress,
-  buyAssetTradeFeeUsd,
-}) => {
-  try {
-    const { assetReference, assetNamespace } = fromAssetId(sellAsset.assetId)
-    const isErc20Trade = assetNamespace === 'erc20'
-    const inboundAddress = await getInboundAddressDataForChain(deps.daemonUrl, sellAsset.assetId)
-    const router = inboundAddress?.router
-    const vault = inboundAddress?.address
-    if (!inboundAddress || !router || !vault)
-      throw new SwapError(`[getPriceRatio]: inboundAddress not found for ETH`, {
-        code: SwapErrorType.RESPONSE_ERROR,
-        details: { inboundAddress },
-      })
-
-    const limit = await getLimit({
-      buyAssetId: buyAsset.assetId,
-      sellAmountCryptoBaseUnit,
-      sellAsset,
-      slippageTolerance,
-      deps,
-      buyAssetTradeFeeUsd,
-    })
-
-    const memo = makeSwapMemo({
-      buyAssetId: buyAsset.assetId,
-      destinationAddress,
-      limit,
-    })
-    const data = await deposit(
-      router,
-      vault,
-      isErc20Trade ? assetReference : '0x0000000000000000000000000000000000000000',
-      sellAmountCryptoBaseUnit,
-      memo,
-    )
-
-    return { data, router }
-  } catch (e) {
-    if (e instanceof SwapError) throw e
-    throw new SwapError('[getThorTxInfo]', { cause: e, code: SwapErrorType.TRADE_QUOTE_FAILED })
-  }
-}
